+++ asistente/auxiliar/checks.py
@@ -1,13 +1,21 @@
 "Módulo para checks auxiliares."
 
-from discord import Interaction, Thread
+from typing import TYPE_CHECKING
+
+from discord import Thread
 from discord.app_commands import check as appcheck
 
 from ..db.atajos import get_admins_por_nivel_y_guild
 from ..db.enums import NivelPermisos
 
+if TYPE_CHECKING:
+    from discord import Interaction
 
-def _verificar_permisos(interaccion: Interaction,
+GUILD_FUNDAMENTOS: int = 653341065767550976
+"ID del servidor de Fundamentos de Programación."
+
+
+def _verificar_permisos(interaccion: "Interaction",
                         nivel: NivelPermisos,
                         al_menos: bool) -> bool:
     if interaccion.guild is None:
@@ -23,14 +31,14 @@ def _verificar_permisos(interaccion: Interaction,
             or any(role.id in candidatos_roles for role in interaccion.user.roles))
 
 
-def tiene_permisos(interaccion: Interaction) -> bool:
+def tiene_permisos(interaccion: "Interaction") -> bool:
     "Verifica si el invocador del comando tiene cualquier clase de privilegios relevantes."
 
     return _verificar_permisos(interaccion, NivelPermisos.PLEBEYO, al_menos=True)
 
 
 def _permisos_check(nivel: NivelPermisos, *, al_menos: bool):
-    def verificar_permisos_predicado(interaccion: Interaction) -> bool:
+    def verificar_permisos_predicado(interaccion: "Interaction") -> bool:
         """
         Verifica si el usuario que invoca el comando tiene los privilegios adecuados.
 
@@ -58,7 +66,22 @@ def permisos_de_al_menos_nivel(nivel: NivelPermisos):
 def es_hilo():
     "Verifica si la interacción ocurrió en un hilo."
 
-    def predicado(interaccion: Interaction) -> bool:
-        return isinstance(interaccion.channel, Thread)
+    def predicado(interaccion: "Interaction") -> bool:
+        return interaccion.channel is not None and isinstance(interaccion.channel, Thread)
 
     return appcheck(predicado)
+
+
+def es_servidor(guild_id: int):
+    "Verifica si la interacción ocurre en un guild dado."
+
+    def predicado(interaccion: "Interaction") -> bool:
+        return (interaccion.guild_id is not None) and (interaccion.guild_id == guild_id)
+
+    return appcheck(predicado)
+
+
+def es_servidor_fundamentos():
+    "Verifica si es específicamente el servidor de la materia 'Fundamentos de Programación'."
+
+    return es_servidor(GUILD_FUNDAMENTOS)