+++ asistente/db/atajos/__init__.py
@@ -5,4 +5,5 @@ hacen la de atajos para db.
 
 from .guias import *
 from .guilds import *
+from .padrones import *
 from .permisos import *