+++ asistente/interfaces/__init__.py
@@ -1,5 +1,8 @@
 "Paquete de interfaces."
 
+# Por diseño se prefiere que se importe explícitamente
+# from .fundamentos import *
+
 from .ahorcado import *
 from .ejs import *
 from .ui_general import *